Implemented Player Movement and Camera Follow
I added a free asset pack from itch.io "https://smallscaleint.itch.io/hd-8-directional-top-down-character-pack-1" that contained Sprite sheets for a knight that will be the player.

I made animations of the spritesheets for idle and walking in all 8 directions. Created an animation controller with blend trees so that the animations smoothly transitions between eachother.

I created 2 scripts, IsoPlayerController and PlayerAnimationController. For updating movement with animations.

Added a Chinemachine Camera that follows the player, also added collider2D and rigidbody2D to the player.