Added a skeleton and downloaded AI Navigation. Updated Camera to match pixels per unit
Added a pixel perfect camera script to the camera for perfect pixel quality on the assets i added to the game, it was hard finding assets with the same pixel quality 128 x 128 is what i was originally going for but i got some other assets but seeing as I'm not an artist and i only will be working on this 3 days i think it does not matter that much.

Added AI Navigation plugin and a skeleton enemy asset pack from unity asset store.