Animation bug fixes, health script created and attack logic added.
Animation error fixes, removed AnimationEventHandler() events that the skeleton animation pack was using that got called in the animations.

Added a running attack to the player, now it has a still and moving attack.

Created a health script and added attack, damage, death functions to both player and enemy. The health script connects the animation logic by calling either player or enemy takedamage or death animations. Depending if "isplayer" bool is active on the script.

Still need to work on the attacks they are not so pretty.