Created a randomized isometric grid/tilemap
Got the tile sprites from itch.io : https://iocainesmoothie.itch.io/free-iso-tiles-grassstonesand

randomly spawned tiles with TileSpawner script and randomly spawn trees on the tiles but not near the player at start with TilemapObjectSpawner script

Got the trees from : https://opengameart.org/content/free-isometric-plants-pack

problem is that the plant pack's sprites are 64 x 32, so the pixel count does not match the rest but i could not find free assets where they did.
This results in the trees being a bit blurry in the game.