Updated trees and created SortingOrder Script
Updated trees colliders to make a better navmesh for AI and player to navigate.

SortingOrder script sets the sorting order on the sprite renderer to the position in y so that when player for example go behind a tree it will be rendered behind it and when walking infront of it the player will render in front.